删除web_chat,自动融合
