+++ server/api.py
@@ -14,13 +14,12 @@ from fastapi.middleware.cors import CORSMiddleware
 from starlette.responses import RedirectResponse
 from server.chat.chat import chat
 from server.chat.openai_chat import openai_chat
-from server.chat.search_engine_chat import search_engine_chat
 from server.chat.completion import completion
 from server.chat.feedback import chat_feedback
 from server.embeddings_api import embed_texts_endpoint
 from server.llm_api import (list_running_models, list_config_models,
                             change_llm_model, stop_llm_model,
-                            get_model_config, list_search_engines)
+                            get_model_config)
 from server.utils import (BaseResponse, ListResponse, FastAPI, MakeFastAPIOffline,
                           get_server_configs, get_prompt_template)
 from typing import List, Literal
@@ -69,11 +68,6 @@ def mount_app_routes(app: FastAPI, run_mode: str = None):
              summary="与llm模型对话(通过LLMChain)",
              )(chat)
 
-    app.post("/chat/search_engine_chat",
-             tags=["Chat"],
-             summary="与搜索引擎对话",
-             )(search_engine_chat)
-
     app.post("/chat/feedback",
              tags=["Chat"],
              summary="返回llm模型对话评分",
@@ -116,16 +110,12 @@ def mount_app_routes(app: FastAPI, run_mode: str = None):
              summary="获取服务器原始配置信息",
              )(get_server_configs)
 
-    app.post("/server/list_search_engines",
-             tags=["Server State"],
-             summary="获取服务器支持的搜索引擎",
-             )(list_search_engines)
 
     @app.post("/server/get_prompt_template",
              tags=["Server State"],
              summary="获取服务区配置的 prompt 模板")
     def get_server_prompt_template(
-        type: Literal["llm_chat", "knowledge_base_chat", "search_engine_chat", "agent_chat"]=Body("llm_chat", description="模板类型，可选值：llm_chat，knowledge_base_chat，search_engine_chat，agent_chat"),
+        type: Literal["llm_chat", "knowledge_base_chat"]=Body("llm_chat", description="模板类型，可选值：llm_chat，knowledge_base_chat"),
         name: str = Body("default", description="模板名称"),
     ) -> str:
         return get_prompt_template(type=type, name=name)
