+++ tests/api/test_server_state_api.py
@@ -29,14 +29,6 @@ def test_server_configs():
     assert len(configs) > 0
 
 
-def test_list_search_engines():
-    engines = api.list_search_engines()
-    pprint(engines)
-
-    assert isinstance(engines, list)
-    assert len(engines) > 0
-
-
 @pytest.mark.parametrize("type", ["llm_chat"])
 def test_get_prompt_template(type):
     print(f"prompt template for: {type}")
