+++ tests/api/test_stream_chat_api.py
@@ -97,29 +97,3 @@ def test_knowledge_chat(api="/chat/knowledge_base_chat"):
     assert "docs" in data and len(data["docs"]) > 0
     assert response.status_code == 200
 
-
-def test_search_engine_chat(api="/chat/search_engine_chat"):
-    global data
-
-    data["query"] = "室温超导最新进展是什么样？"
-
-    url = f"{api_base_url}{api}"
-    for se in ["bing", "duckduckgo"]:
-        data["search_engine_name"] = se
-        dump_input(data, api + f" by {se}")
-        response = requests.post(url, json=data, stream=True)
-        if se == "bing" and not BING_SUBSCRIPTION_KEY:
-            data = response.json()
-            assert data["code"] == 404
-            assert data["msg"] == f"要使用Bing搜索引擎，需要设置 `BING_SUBSCRIPTION_KEY`"
-
-        print("\n")
-        print("=" * 30 + api + f" by {se}  output" + "="*30)
-        for line in response.iter_content(None, decode_unicode=True):
-            data = json.loads(line)
-            if "answer" in data:
-                print(data["answer"], end="", flush=True)
-        assert "docs" in data and len(data["docs"]) > 0
-        pprint(data["docs"])
-        assert response.status_code == 200
-
